requirements.txt
Django~=5.2.5
django_autocomplete_light==3.12.1
django_environ==0.12.0
djangorestframework_simplejwt==5.5.1
openpyxl==3.1.5
pricing==1.0.1
resources==0.0.1
tablib==3.8.0
typing_extensions==4.15.0
weasyprint==68.1

djangorestframework~=3.16.1
typing_extensions~=4.14.1
django-filter~=25.1
django-debug-toolbar~=6.0.0